MachineLearning/PyTorch/04_VisualizeFeatureMap/requirements.txt
absl-py==1.1.0
art==5.6
cachetools==4.2.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
colorlog==6.6.0
cycler==0.11.0
dataclasses==0.8
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.2.0
ordereddict==1.1
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycm==3.5
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
torch==1.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
zipp==3.6.0
